Move tests from Program to a dedicated class
